extract locale from cookie
