+++ inlang/packages/paraglide/paraglide-js/src/compiler/runtime/get-locale.js
@@ -1,6 +1,6 @@
 import { assertIsLocale } from "./assert-is-locale.js";
 import { baseLocale } from "./base-locale.js";
-import { cookieName } from "./cookie-name.js";
+import { extractLocaleFromCookie } from "./extract-locale-from-cookie.js";
 import { extractLocaleFromPathname } from "./extract-locale-from-pathname.js";
 import { strategy } from "./strategy.js";
 
@@ -56,7 +56,3 @@ export let getLocale = () => {
 	throw new Error("No locale found. There is an error in your strategy.");
 };
 
-function extractLocaleFromCookie() {
-	const match = document.cookie.match(new RegExp(`(^| )${cookieName}=([^;]+)`));
-	return match?.[2];
-}
