+++ inlang/packages/paraglide/paraglide-js/src/compiler/runtime/get-locale.test.ts
@@ -1,28 +1,7 @@
 import { test, expect } from "vitest";
 import { createRuntimeForTesting } from "./create-runtime.js";
 
-// @ts-expect-error - global variable definition
-globalThis.document = {};
-
-test("matches the locale of a cookie", async () => {
-	const runtime = await createRuntimeForTesting({
-		baseLocale: "en",
-		locales: ["en", "de"],
-		compilerOptions: {
-			isServer: "false",
-			strategy: ["cookie"],
-			cookieName: "PARAGLIDE_LOCALE",
-		},
-	});
-
-	globalThis.document.cookie =
-		"OTHER_COOKIE=fr; PARAGLIDE_LOCALE=de; ANOTHER_COOKIE=en; EXPIRES_COOKIE=es; Max-Age=3600";
-
-	const locale = runtime.getLocale();
-	expect(locale).toBe("de");
-});
-
-test("falls back to base locale if no cookie is matched", async () => {
+test("matching by strategy works", async () => {
 	const baseLocale = "en";
 
 	const runtime = await createRuntimeForTesting({
@@ -35,6 +14,8 @@ test("falls back to base locale if no cookie is matched", async () => {
 		},
 	});
 
+	// @ts-expect-error - global variable definition
+	globalThis.document = {};
 	globalThis.document.cookie = "OTHER_COOKIE=blaba;";
 
 	const locale = runtime.getLocale();
